+++ pyproject.toml
@@ -11,6 +11,8 @@ fastapi = "^0.108.0"
 pydantic = "^2.5.3"
 sqlalchemy = "^2.0.25"
 uvicorn = "^0.25.0"
+setuptools = "^69.0.3"
+httpx = "^0.26.0"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.12.1"
