+++ src/data/dummy.py
@@ -1,0 +1,6 @@
+from typing import List
+def get_one(name: str) -> Dummy:
+    return Dummy(name)
+
+def get_all() -> List[Dummy]:
+    ...
